refactor: split balance chart components
