+++ frontend/src/accounts/detail/AccountDetailPage.tsx
@@ -7,7 +7,7 @@ import type { Transaction } from '@/api/transactions'
 import CreateTransactionModal from '@/components/CreateTransactionModal'
 import AccountIdentityCard from '@/accounts/detail/components/AccountIdentityCard'
 import BackLink from '@/accounts/detail/components/BackLink'
-import BalanceChartCard from '@/accounts/detail/components/BalanceChartCard'
+import BalanceChartCard from '@/accounts/detail/components/balance-chart/BalanceChartCard'
 import EditAccountIdentityModal from '@/accounts/detail/components/EditAccountIdentityModal'
 import MonthlyCashFlowCard from '@/accounts/detail/components/MonthlyCashFlowCard'
 import { TopCategoriesBySpendingCard, TopMerchantsBySpendingCard } from '@/accounts/detail/components/SpendingBreakdownCards'
